RDKCOM-2632: REFPLTB-1383 : Need to resolve build time warning in "ccsp-tr06...
Reason for change: Adding 64bit support
Test Procedure: Built 32 bit and 64 bit image
Risks: Low

(cherry picked from commit 2a353087389f8967efb9e785a317a5ebcd4ebfe8)

+++ source-embedded/CcspCwmpTcpConnReqHandler/ccsp_cwmp_tcpcrho_auth.c
@@ -1987,10 +1987,10 @@ CcspCwmpTcpcrhoGenResponse
         (
             pMsg+AnscSizeOfString(pMsg),
             ulSize - AnscSizeOfString(pMsg),
-            "Set-Cookie: %s=\"%u/%u\"; Version=\"1\"; Path=\"/\"\r\n",
+            "Set-Cookie: %s=\"%u/%lu\"; Version=\"1\"; Path=\"/\"\r\n",
             CCSP_CWMP_TCPCR_COOKIE_MAGIC_NUMBER,
             (unsigned int)pMyObject->MagicID,
-            (unsigned int)buffer
+            (long unsigned)buffer
         );
 
     /* WWW-Authenticate */
